src/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-analysis-system"
version = "0.1.0"
description = "A financial data analysis system with natural language processing capabilities"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    # 核心依赖
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=0.19.0",
    
    # LLM & AI 相关
    "openai>=1.0.0",
    "litellm>=0.1.0",
    "pandasai>=0.1.0",
    "smol-agent>=0.1.0",
    
    # 数据可视化
    "matplotlib>=3.5.0",
    "plotly>=5.5.0",
    "seaborn>=0.11.0",
    
    # Web 界面
    "streamlit>=1.15.0",
    
    # 工具和辅助库
    "typing-extensions>=4.0.0",
    "pydantic>=1.9.0",
    "loguru>=0.6.0",
    "python-json-logger>=2.0.0",
    "tqdm>=4.64.0",
    
    # 数据处理和分析
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src"
testpaths = [
    "tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true
